added

+++ src/main/java/com/pharmacy/management/dto/request/ProductRequestDTO.java
@@ -12,18 +12,14 @@ public class ProductRequestDTO {
     private String name;
     @NotNull
     private String code;
-//    @NotNull
-//    private String productId;
     private String description;
     private String strength;
     private String stsMedicine;
     @NotNull
     private Long categoryId;
     private Long supplierId;
-//    @Enumerated(EnumType.STRING)
-//    private OrderType orderType;
     @NotNull
-    private Integer unitPrice;
+    private Double unitPrice;
     private Integer limitCost;
     private Integer limitUnit;
     private byte[] file;
